games/loggingservice/requirements.txt
fastapi
uvicorn[standard]
pydantic
asyncpg
python-dotenv

